Fix bugs when merge the latest develop branch.
