+++ components/pages/AppPages/Today.js
@@ -8,7 +8,7 @@ import styles from "../../../styles/scss/application.module.scss";
 
 function Today(props) {
   const router = useRouter();
-  localStorage.setItem("lastPage", "today");
+  // localStorage.setItem("lastPage", "today");
 
   //use timestamp to compare if the item dueDate is today
   const todayTasks = props.data.filter((task) => {
